Cleaned up a bit. Ready for production.
